Adding add medicine for caregiver

+++ mobile/app/src/main/java/fr/android/mhealthy/model/Patient.java
@@ -5,10 +5,10 @@ import java.io.Serializable;
 public class Patient implements Serializable {
     public int id;
     public String name;
-    public int add_time;
+    public long add_time;
     public String phone;
 
-    public Patient(int id, String name, int time, String phone) {
+    public Patient(int id, String name, long time, String phone) {
         this.id = id;
         this.name = name;
         this.add_time = time;
